+++ Kernel/System/CustomerUser.pm
@@ -733,22 +733,10 @@ sub CustomerName {
 
         next SOURCE if !$Self->{"CustomerUser$Count"};
 
-        # get customer name and return it
-        my %Name = $Self->{"CustomerUser$Count"}->CustomerUserDataGet( User => $Param{UserLogin} );
-
-        # get configuration for the full name order
-        my $FirstnameLastNameOrder = $Kernel::OM->Get('Kernel::Config')->Get('FirstnameLastnameOrder') || 0;
-
-        if (%Name) {
-
-            # generate the full name and save it in the hash
-            my $UserFullname = $Self->_CustomerUserFullname(
-                UserFirstname => $Name{UserFirstname},
-                UserLastname  => $Name{UserLastname},
-                UserLogin     => $Name{UserLogin},
-                NameOrder     => $FirstnameLastNameOrder,
-            );
-            return $UserFullname;
+        # Get customer name and return it.
+        my $Name = $Self->{"CustomerUser$Count"}->CustomerName(%Param);
+        if ($Name) {
+            return $Name;
         }
     }
     return;
@@ -839,9 +827,6 @@ sub CustomerUserDataGet {
     my $CustomerCompanyObject     = $Kernel::OM->Get('Kernel::System::CustomerCompany');
     my $DynamicFieldBackendObject = $Kernel::OM->Get('Kernel::System::DynamicField::Backend');
 
-    # get configuration for the full name order
-    my $FirstnameLastNameOrder = $ConfigObject->Get('FirstnameLastnameOrder') || 0;
-
     SOURCE:
     for my $Count ( '', 1 .. 10 ) {
 
@@ -851,10 +836,7 @@ sub CustomerUserDataGet {
         next SOURCE if !%Customer;
 
         # generate the full name and save it in the hash
-        my $UserFullname = $Self->_CustomerUserFullname(
-            %Customer,
-            NameOrder => $FirstnameLastNameOrder,
-        );
+        my $UserFullname = $Self->CustomerName(%Customer);
 
         # save the generated fullname in the hash.
         $Customer{UserFullname} = $UserFullname;
@@ -1350,86 +1332,6 @@ sub CustomerUserCacheClear {
 
     return 1;
 }
-
-=begin Internal:
-
-=head2 _CustomerUserFullname()
-
-Builds the customer user full name based on first name, last name. The order
-can be configured.
-
-    my $Fullname = $Object->_CustomerUserFullname(
-        UserFirstname => 'Test',
-        UserLastname  => 'Person',
-        UserLogin     => 'tp',
-        NameOrder     => 0,         # optional 0, 1, 2, 3, 4, 5
-    );
-
-=cut
-
-sub _CustomerUserFullname {
-    my ( $Self, %Param ) = @_;
-
-    # For databases which do not have configured CustomerUser information use empty string
-    #   for full name creation. See bug#13587 for more information.
-    for my $Data (qw(UserFirstname UserLastname UserLogin)) {
-        $Param{$Data} //= '';
-    }
-
-    my $FirstnameLastNameOrder = $Param{NameOrder} || 0;
-
-    my $UserFullname;
-    if ( $FirstnameLastNameOrder eq '0' ) {
-        $UserFullname = $Param{UserFirstname} . ' '
-            . $Param{UserLastname};
-    }
-    elsif ( $FirstnameLastNameOrder eq '1' ) {
-        $UserFullname = $Param{UserLastname} . ', '
-            . $Param{UserFirstname};
-    }
-    elsif ( $FirstnameLastNameOrder eq '2' ) {
-        $UserFullname = $Param{UserFirstname} . ' '
-            . $Param{UserLastname} . ' ('
-            . $Param{UserLogin} . ')';
-    }
-    elsif ( $FirstnameLastNameOrder eq '3' ) {
-        $UserFullname = $Param{UserLastname} . ', '
-            . $Param{UserFirstname} . ' ('
-            . $Param{UserLogin} . ')';
-    }
-    elsif ( $FirstnameLastNameOrder eq '4' ) {
-        $UserFullname = '(' . $Param{UserLogin}
-            . ') ' . $Param{UserFirstname}
-            . ' ' . $Param{UserLastname};
-    }
-    elsif ( $FirstnameLastNameOrder eq '5' ) {
-        $UserFullname = '(' . $Param{UserLogin}
-            . ') ' . $Param{UserLastname}
-            . ', ' . $Param{UserFirstname};
-    }
-    elsif ( $FirstnameLastNameOrder eq '6' ) {
-        $UserFullname = $Param{UserLastname} . ' '
-            . $Param{UserFirstname};
-    }
-    elsif ( $FirstnameLastNameOrder eq '7' ) {
-        $UserFullname = $Param{UserLastname} . ' '
-            . $Param{UserFirstname} . ' ('
-            . $Param{UserLogin} . ')';
-    }
-    elsif ( $FirstnameLastNameOrder eq '8' ) {
-        $UserFullname = '(' . $Param{UserLogin}
-            . ') ' . $Param{UserLastname}
-            . ' ' . $Param{UserFirstname};
-    }
-    elsif ( $FirstnameLastNameOrder eq '9' ) {
-        $UserFullname = $Param{UserLastname} . $Param{UserFirstname};
-    }
-    return $UserFullname;
-}
-
-=end Internal:
-
-=cut
 
 =head2 CustomerUserCustomerMemberAdd()